requirements.txt
amqp==2.5.2
apns2==0.7.1
autopep8==1.3.5
Babel==2.6.0
beautifulsoup4==4.9.0
billiard==3.6.2.0
celery==4.4.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
cycler==0.10.0
Django==2.2
django-adminactions==1.8.1
django-authtools==1.6.0
django-background-tasks==1.2.5
django-bootstrap4==1.0.1
django-braces==1.13.0
django-colorfield==0.3.2
django-compat==1.0.15
django-crispy-forms==1.7.2
django-crontab==0.7.1
django-debug-toolbar==1.10.1
django-environ==0.4.5
django-filter==2.2.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-formtools==2.1
django-phonenumber-field==2.0.1
django-push-notifications==2.0.0
django-rest-auth==0.9.5
django-tables2==2.1.0
djangorestframework==3.11.0
easy-thumbnails==2.5
et-xmlfile==1.0.1
future==0.18.2
get==2018.11.19
googlemaps==4.3.1
h2==2.6.2
hpack==3.0.0
html5lib==1.0.1
http-ece==1.1.0
hyper==0.7.0
hyperframe==3.2.0
idna==2.8
importlib-metadata==1.5.0
jdcal==1.4.1
jsonfield==3.1.0
kiwisolver==1.1.0
kombu==4.6.7
matplotlib==3.1.2
ndg-httpsclient==0.5.1
numpy==1.17.4
oauthlib==3.1.0
openpyxl==3.0.3
phonenumbers==8.9.12
Pillow==5.3.0
post==2018.11.20
protobuf==3.9.2
psycopg2==2.8.5
psycopg2-binary==2.7.5
public==2018.11.20
pusher==3.0.0
py-vapid==1.7.0
pyasn1==0.4.8
pycodestyle==2.4.0
pycparser==2.20
PyJWT==1.7.1
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyparsing==2.4.5
PyPDF2==1.26.0
python-dateutil==2.8.1
python-decouple==3.3
python-http-client==3.2.5
python-twitter==3.5
pytz==2018.7
pywebpush==1.10.1
query-string==2018.11.20
ratelimit==2.2.1
reportlab==3.5.42
request==2018.11.20
requests==2.22.0
requests-oauthlib==1.3.0
selenium==3.141.0
sendgrid==6.1.2
six==1.14.0
smartystreets-python-sdk==4.5.0
soupsieve==2.0
sqlparse==0.2.4
unicodecsv==0.14.1
urllib3==1.25.7
vine==1.3.0
virtualenv==16.6.0
webencodings==0.5.1
Werkzeug==0.14.1
xhtml2pdf==0.2.4
xlrd==1.2.0
xlwt==1.3.0
zipp==2.2.0
